Refactor copy button, creating component #73
